backend/requirements.txt
Django==4.2
django-storages[azure]==1.13.2
azure-storage-blob==12.16.0
django-cors-headers==4.2.0
djangorestframework==3.14.0
python-dotenv==1.0.0
whitenoise==6.5.0
psycopg2-binary==2.9.7